pyproject.toml
[project]
name = "test-sites"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
]
